pyproject.toml
[tool.poetry]
name = "lucas"
version = "0.1.0"
description = "Monitora portais da transparência de acordo com a LAI e LRF"
authors = ["Ana Paula Gomes <apgomes88@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
lucas = "lucas.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
bs4 = "^0.0.1"
ipython = "^8.12.0"
pytest-playwright = "^0.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
requests-mock = "^1.10.0"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "rules: marca um teste como uma regra (pule-os com '-m \"not rules\"')",
]
